bs-code-master/semseg/requirements.txt
Pillow
torch
torchvision
numpy
tqdm
opencv-python
